requirements.txt
TgCrypto
https://github.com/CW4RR10R/pyrogram/archive/asyncio.zip
git+https://github.com/ytdl-org/youtube-dl
Pillow
hachoir